+++ icekit/page_types/article/migrations/0006_auto_20161117_1800.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('icekit_article', '0005_add_hero'),
+    ]
+
+    operations = [
+        migrations.AlterModelOptions(
+            name='articlecategorypage',
+            options={},
+        ),
+    ]