+++ icekit/page_types/layout_page/abstract_models.py
@@ -1,7 +1,7 @@
 from fluent_pages.integration.fluent_contents import FluentContentsPage
 
 from icekit.publishing.models import PublishableFluentContentsPage
-from icekit.mixins import LayoutFieldMixin
+from icekit.mixins import LayoutFieldMixin, HeroMixin, ListableMixin
 
 
 class AbstractUnpublishableLayoutPage(FluentContentsPage, LayoutFieldMixin):
@@ -9,6 +9,10 @@ class AbstractUnpublishableLayoutPage(FluentContentsPage, LayoutFieldMixin):
         abstract = True
 
 
-class AbstractLayoutPage(PublishableFluentContentsPage, LayoutFieldMixin):
+class AbstractLayoutPage(PublishableFluentContentsPage, LayoutFieldMixin, HeroMixin, ListableMixin):
     class Meta:
         abstract = True
+
+    def get_type(self):
+        # we don't normally want pages to say they're a 'page'
+        return ""