+++ icekit/page_types/layout_page/models.py
@@ -1,6 +1,5 @@
 from . import abstract_models
 
-
 class LayoutPage(abstract_models.AbstractLayoutPage):
 
     class Meta: