+++ icekit/project/settings/_base.py
@@ -30,15 +30,16 @@ REDIS_ADDRESS = os.environ.get('REDIS_ADDRESS', 'localhost:6379')
 
 # Uniquely identify the base settings module, so we can avoid conflicts with
 # other projects running on the same system.
-SETTINGS_MODULE_HASH = hashlib.md5(__file__ + BASE_SETTINGS_MODULE).hexdigest()
+SETTINGS_MODULE_HASH = hashlib.md5(
+    u''.join((__file__, BASE_SETTINGS_MODULE)).encode('utf-8')).hexdigest()
 
 PROJECT_NAME = os.environ.get('ICEKIT_PROJECT_NAME', 'ICEkit')
-PROJECT_SLUG = re.sub(r'[^0-9A-Za-z]+', '-', slugify(unicode(PROJECT_NAME)))
+PROJECT_SLUG = re.sub(r'[^0-9A-Za-z]+', '-', slugify(PROJECT_NAME))
 
 SITE_DOMAIN = os.environ.get('SITE_DOMAIN', '%s.lvh.me' % PROJECT_SLUG)
 SITE_NAME = os.environ.get('SITE_NAME', PROJECT_NAME)
 
-SITE_PORT = 8000
+SITE_PORT = None
 
 # FILE SYSTEM PATHS ###########################################################
 
@@ -202,6 +203,7 @@ INSTALLED_APPS = (
     'forms_builder.forms',
     # 'ixc_redactor',
     'reversion',
+    'django_object_actions',
 
     # Default.
     'django.contrib.admin',
@@ -342,7 +344,7 @@ SITE_ID = 1
 
 # CELERY ######################################################################
 
-BROKER_URL = CELERY_RESULT_BACKEND = 'redis://%s/0' % REDIS_ADDRESS
+BROKER_URL = 'redis://%s/0' % REDIS_ADDRESS
 CELERY_ACCEPT_CONTENT = ['json', 'msgpack', 'yaml']  # 'pickle'
 CELERY_DEFAULT_QUEUE = PROJECT_SLUG
 CELERY_TASK_SERIALIZER = 'json'
@@ -366,6 +368,7 @@ CELERYBEAT_SCHEDULE = {
     },
 }
 
+# Redis (by setting CELERY_RESULT_BACKEND to BROKER_URL) is an alternative
 CELERY_RESULT_BACKEND = 'djcelery.backends.database:DatabaseBackend'
 
 # Log the celerybeat lock actions
@@ -397,13 +400,13 @@ COMPRESS_OFFLINE_CONTEXT = 'ixc_compressor.get_compress_offline_context'
 COMPRESS_PRECOMPILERS = (
     (
         'text/less',
-        '%s {infile} {outfile} --autoprefix' % (
+        '"%s" {infile} {outfile} --autoprefix' % (
             os.path.join(PROJECT_DIR, 'node_modules', '.bin', 'lessc'),
         ),
     ),
     (
         'text/x-scss',
-        '%s {infile} {outfile} --autoprefix --include-path %s' % (
+        '"%s" {infile} {outfile} --autoprefix --include-path %s' % (
             os.path.join(PROJECT_DIR, 'node_modules', '.bin', 'node-sass'),
             STATIC_ROOT,
         ),
@@ -453,7 +456,10 @@ THUMBNAIL_ALIASES = {
         },
         'image_gallery_thumb': {
             'size': (200, 0),
-        }
+        },
+        'list_image': {
+            'size': (150, 0),
+        },
     }
 }
 
@@ -473,26 +479,27 @@ INSTALLED_APPS += ('flat', )
 DJANGO_WYSIWYG_FLAVOR = 'alloyeditor'
 DJANGO_WYSIWYG_MEDIA_URL = STATIC_URL + 'alloyeditor/dist/alloy-editor/'
 
-_BASIC_PLUGINS = [
+BASIC_PLUGINS = [
     'RawHtmlPlugin',
     'TextPlugin',
     'HorizontalRulePlugin',
 ]
 
-_TEXT_PLUGINS = [
+TEXT_PLUGINS = [
     'FAQPlugin',
     'QuotePlugin',
 ]
 
-_ASSETS_PLUGINS = [
+ASSETS_PLUGINS = [
     'SlideShowPlugin',
     'ImagePlugin',
     'ImageGalleryPlugin',
     'FilePlugin',
     'SharedContentPlugin',
+    'ContactPersonPlugin',
 ]
 
-_EMBED_PLUGINS = [
+EMBED_PLUGINS = [
     'IframePlugin',
     'MapPlugin',
     'MapWithTextPlugin',
@@ -503,26 +510,33 @@ _EMBED_PLUGINS = [
     'TwitterEmbedPlugin',
 ]
 
-_NAVIGATION_PLUGINS = [
+NAVIGATION_PLUGINS = [
     'PageAnchorPlugin',
     'PageAnchorListPlugin',
     'ChildPagesPlugin',
 ]
 
+LINK_PLUGINS = [
+    'ArticleLinkPlugin',
+    'PageLinkPlugin',
+    'AuthorLinkPlugin',
+]
+
 DEFAULT_PLUGINS = \
-    _BASIC_PLUGINS + \
-    _TEXT_PLUGINS + \
-    _ASSETS_PLUGINS + \
-    _EMBED_PLUGINS + \
-    _NAVIGATION_PLUGINS
+    BASIC_PLUGINS + \
+    TEXT_PLUGINS + \
+    ASSETS_PLUGINS + \
+    EMBED_PLUGINS + \
+    NAVIGATION_PLUGINS + \
+    LINK_PLUGINS
 
 FLUENT_CONTENTS_PLACEHOLDER_CONFIG = {
     'main': {
         'plugins': DEFAULT_PLUGINS,
     },
-    # 'sidebar': {
-    #     'plugins': ('...', ),
-    # },
+    'related': {
+        'plugins': LINK_PLUGINS,
+    }
 }
 
 FLUENT_DASHBOARD_DEFAULT_MODULE = 'ModelList'
@@ -575,7 +589,7 @@ INSTALLED_APPS += (
     # 'fluent_contents.plugins.picture',
     'fluent_contents.plugins.rawhtml',
     'fluent_contents.plugins.sharedcontent',
-    'fluent_contents.plugins.text',
+    # 'fluent_contents.plugins.text',
     # 'fluent_contents.plugins.twitterfeed',
 
     # Page type and content plugin dependencies.
@@ -637,7 +651,7 @@ ICEKIT = {
         ),
     ),
 
-    'DASHBOARD_FEATURED_APPS': (
+    'DASHBOARD_FEATURED_APPS': [
         {
             'verbose_name': 'Content',
             'icon_html': '<i class="content-type-icon fa fa-files-o"></i>',
@@ -660,7 +674,7 @@ ICEKIT = {
                 # 'sharedcontent.SharedContent': {},
             },
         },
-    ),
+    ],
 }
 
 INSTALLED_APPS += (
@@ -680,11 +694,13 @@ INSTALLED_APPS += (
 
     # 'icekit.plugins.brightcove',
     'icekit.plugins.child_pages',
+    'icekit.plugins.contact_person',
     'icekit.plugins.faq',
     'icekit.plugins.file',
     'icekit.plugins.horizontal_rule',
     'icekit.plugins.image',
     'icekit.plugins.instagram_embed',
+    'icekit.plugins.links',
     'icekit.plugins.map',
     'icekit.plugins.map_with_text',
     'icekit.plugins.oembed_with_caption',
@@ -695,6 +711,7 @@ INSTALLED_APPS += (
     'icekit.plugins.slideshow',
     'icekit.plugins.image_gallery',
     'icekit.plugins.twitter_embed',
+    'icekit.plugins.text',
 )
 
 MIDDLEWARE_CLASSES += ('icekit.publishing.middleware.PublishingMiddleware', )