+++ icekit/project/settings/_develop.py
@@ -1,13 +1,16 @@
 from ._base import *
 
+SITE_PORT = 8000
+
 # DJANGO ######################################################################
 
 ALLOWED_HOSTS = ('*', )
 
-CACHES['default'].update({
-    'BACKEND': 'redis_lock.django_cache.RedisCache',
-    'LOCATION': 'redis://%s/1' % REDIS_ADDRESS,
-})
+# Disabling develop cache backend as short-term workaround until #149 is resolved
+# CACHES['default'].update({
+#     'BACKEND': 'redis_lock.django_cache.RedisCache',
+#     'LOCATION': 'redis://%s/1' % REDIS_ADDRESS,
+# })
 
 CSRF_COOKIE_SECURE = False  # Don't require HTTPS for CSRF cookie
 SESSION_COOKIE_SECURE = False  # Don't require HTTPS for session cookie