+++ project_template/go.sh
@@ -23,13 +23,13 @@ if [[ ! -d "$ICEKIT_VENV" ]]; then
     virtualenv "$ICEKIT_VENV"
 fi
 
-# Install ICEKit.
-if [[ -z $("$ICEKIT_VENV/bin/pip" freeze | grep django-icekit) ]]; then
-    "$ICEKIT_VENV/bin/pip" install -r requirements-icekit.txt
+# Install ICEKit project.
+if [[ -z $("$ICEKIT_VENV/bin/python" -m pip freeze | grep django-icekit) ]]; then
+    "$ICEKIT_VENV/bin/python" -m pip install -r requirements-icekit.txt
 fi
 
 # Get absolute directory for the `icekit` package.
-export ICEKIT_DIR=$("$ICEKIT_VENV/bin/python" -c 'import icekit, os; print os.path.abspath(os.path.dirname(icekit.__file__));')
+export ICEKIT_DIR=$("$ICEKIT_VENV/bin/python" -c "import icekit, os, sys; sys.stdout.write('%s\n' % os.path.dirname(icekit.__file__));")
 
 # Execute entrypoint and command.
 exec "$ICEKIT_DIR/bin/entrypoint.sh" ${@:-setup-django.sh bash.sh}